+++ mkanak_spring/src/main/java/mkanak_spring/MkanakSpringApplication.java
@@ -1,9 +1,8 @@
 package mkanak_spring;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
+@SpringBootApplication
 public class MkanakSpringApplication {
 
 	public static void main(String[] args) {
