+++ mkanak_spring/src/main/java/mkanak_spring/model/PhoneNumberId.java
@@ -1,24 +1,19 @@
-package mkanak_spring.models;
+package mkanak_spring.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-@Entity
-@Table(name = "phone_numbers")
+import java.io.Serializable;
+@EqualsAndHashCode
+@AllArgsConstructor
 @NoArgsConstructor
-@IdClass(PhoneNumberId.class)
-public class PhoneNumber {
-    @Id
+public class PhoneNumberId implements Serializable {
     @Column(name = "user_id")
     private Long user_id;
-    @Id
     @Column(name = "phone_number")
     private String phone_number;
-
-    public PhoneNumber(Long user_id, String phone_number) {
-        this.user_id = user_id;
-        this.phone_number = phone_number;
-    }
 
     public Long getUser_id() {
         return user_id;
@@ -36,5 +31,3 @@ public class PhoneNumber {
         this.phone_number = phone_number;
     }
 }
-
-
