+++ mkanak_spring/src/main/java/mkanak_spring/model/dao/PhoneNumberRepo.java
@@ -1,7 +1,7 @@
-package mkanak_spring.models.dao;
+package mkanak_spring.model.dao;
 
-import mkanak_spring.models.PhoneNumberId;
-import mkanak_spring.models.PhoneNumber;
+import mkanak_spring.model.PhoneNumberId;
+import mkanak_spring.model.PhoneNumber;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
