+++ mkanak_spring/src/main/java/mkanak_spring/model/dao/UserDAO.java
@@ -1,7 +1,7 @@
-package mkanak_spring.models.dao;
+package mkanak_spring.model.dao;
 
-import mkanak_spring.models.User;
-import mkanak_spring.models.UserCredentials;
+import mkanak_spring.model.User;
+import mkanak_spring.model.UserCredentials;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
