+++ mkanak_spring/src/main/java/mkanak_spring/model/services/UserService.java
@@ -1,9 +1,12 @@
-package mkanak_spring.models.services;
+package mkanak_spring.model.services;
 
+import mkanak_spring.model.User;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.ParseException;
 
 public interface UserService {
     String saveUser(JSONObject user) throws ParseException;
     Long logInUser(String username, String password);
+    // edit user
+    // get user
 }
