--- mkanak_spring/src/main/java/mkanak_spring/models/UserCredentials.java
@@ -1,40 +1,0 @@
-package mkanak_spring.models;
-
-import jakarta.persistence.*;
-import org.hibernate.annotations.Immutable;
-
-@Entity
-@Immutable
-@Table(name = "login_credentials")
-public class UserCredentials {
-    @Id
-    private Long user_id;
-    private String username;
-    private String user_password;
-
-    public Long getUser_id() {
-        return user_id;
-    }
-
-    public void setUser_id(Long user_id) {
-        this.user_id = user_id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getUser_password() {
-        return user_password;
-    }
-
-    public void setUser_password(String user_password) {
-        this.user_password = user_password;
-    }
-}
-
-
